free functions cleaned
